==> 02a_Retinopathy_Data_Exploration.Rmd <==
---
title: "REtinopathy Data Exploration"
author: "Tasos Mangelis" 
date: "22.09.2022" 
output:
  powerpoint_presentation:
    reference_doc: my-styles.pptx
---

```{r setup, include = FALSE}

knitr::opts_chunk$set(collapse = TRUE,comment = "#>")

library("DataExplorer")
library(pander)
library(knitr)
library(summarytools)
library(gtsummary)
library(ggplot2)
library(stringr)

```


# Check missing data

# Summary statistics 

## Check data types and missing value counts
```{r echo=FALSE, message=FALSE, warning=FALSE, results='asis', fig.height=6, fig.width=8}
str(retinopathy_DTI)%>%kable()
```

## Quick data overview

```{r echo=FALSE, message=FALSE, warning=FALSE, results='asis', fig.height=6, fig.width=8}
summary(retinopathy_DTI)%>%kable()
```

## Missing data visual

```{r echo=FALSE, message=FALSE, warning=FALSE, results='markup'}
plot_missing(retinopathy_DTI)
```

## Further visualisation to check for potential correlation
```{r echo=FALSE, message=FALSE, warning=FALSE, results='asis', fig.height=6, fig.width=8}
naniar::vis_miss(retinopathy_DTI%>%sample_frac(0.2))
```

# Imputation check by Density plot of the old distribution against the new
```{r echo=FALSE, message=FALSE, warning=FALSE, results="markup", fig.height=6, fig.width=8}

mice::densityplot(prediction)
```

# Histograms

```{r echo=FALSE, message=FALSE, warning=FALSE, results='asis', fig.height=6, fig.width=8}
plot_histogram(precard_DTI_full)
```

# Check for correlations between all variables with visual representation

```{r echo=FALSE, message=FALSE, warning=FALSE, results='asis', fig.height=6, fig.width=8}
plot_correlation(precard_DTI_full)
```
